refactor: split replay benchmark libraries
